getRtopParams(rtop)	R Documentation

Setting parameters for the intamap package

Description

This function sets a range of the parameters for the intamap package, to be included in the object described in rtopIntroduction

Usage

getRtopParams(oldPar,newPar,...)

Arguments

oldPar 	 An existing set of parameters for the interpolation process, of class IntamapParams or a list of parameters for modification of the default parameters

newPar 	 A list of parameters for updating oldPar or for modification of the default parameters. Possible parameters with their defaults are given below

... 	 Individual parameters for updating oldPar or for modification of the default parameters. Possible parameters with their defaults are given below

model = "Ex1"

variogram model type. Currently only "Ex1" is implemented, which is the same model as used in Skoien et al(2006).

nugget = TRUE

variogram model type. Currently only "Ex1" is implemented, which is the same model as used in Skoien et al(2006).

unc = TRUE

variogram model type. Currently only "Ex1" is implemented, which is the same model as used in Skoien et al(2006).

rresol = 100

minimum number of discretization points in areas

hresol = 5

number of discretization points in one direction for areas in binned variograms

cloud = FALSE

logical; if true use the cloud variogram

amul = 1

defines the number of areal bins within one order of magnitude

dmul = 5

defines the number of distance bins within one order of magnitude

ils = 4

defines the type of Least Square method for fitting of variogram

1

least squares difference - err = yobs-ymod

2

weighted least squares difference according to Cressie (1985) - err2=abs(yobs/ymod-1)

3

opposite of Cressie - value overestimations more: err3 = abs(ymod/yobs-1)

4

neutral WLS-method - err = min(err2,err3)

geoDistEst = FALSE

use geostatistical distance when fitting variograms

geoDistPred = FALSE

use geostatistical distance for semivariogram matrices

geoDist

parameter to set jointly geoDistEst = geoDistPred = geoDist

maxdist = Inf

parameter to set jointly geoDistEst = geoDistPred = geoDist

nmax = 10

for local kriging: the number of nearest observations that should be used for a kriging prediction or simulation, where nearest is defined in terms of the space of the spatial locations. By default, 10 observations are used.

hstype = "regular"

sampling type for binned variograms

rstype = "rtop"

sampling type for real areas, see also rtopDiscAreas

wlim = 1.5

an upper limit for the norm of the weights in kriging, see rtopKrige

cv = FALSE

logical; for cross-validation of observations

debug.level = 1

Used in some functions for giving additional output. See individual functions for more information.

observations

observations, used for initial values of parameters if passed

formulaString

formulaString, used for initial values of parameters if passed

Value

A list of the parameters with class rtopParams to be included in the object described in rtopIntroduction

Note

This function will mainly be called by createRtopObject, but can also be called by the user to create a parameter set or update an existing parameter set. If none of the arguments is a list of class rtopParams, the function will assume that the argument(s) are modifications to the default set of parameters. The function can also be called by other functions in the rtop-package if the users chooses not to work with an object of class rtop.

If the function is called with two lists of parameters (but the first one is not of class rtopParams) they are both seen as modifications to the default parameter set. If they share some parameters, the parameter values from the second list will be applied.

Author(s)

Jon Olav Skoien

References

Cressie, N. 1985. Fitting variogram models by weighted least squares. Mathematical Geology, 17 (5), 563-586

Skoien J. O., R. Merz, and G. Bloschl. Top-kriging - geostatistics on stream networks. Hydrology and Earth System Sciences, 10:277-287, 2006

See Also

createRtopObject and rtopIntroduction

Examples


# Create a new set of intamapParameters, with default parameters:
params = getRtopParams()
# Make modifications to the default list of parameters
params = getRtopParams(newPar=list(geoDist = TRUE,nugget = FALSE))
# Make modifications to an existing list of parameters
params = getRtopParams(oldPar = params,newPar = list(geoDist = TRUE,nugget = FALSE))

[Package rtop version 0.3-3 Index]