readAreaInfo(rtop)	R Documentation

create SpatialPointsDataFrame with observations of data with a spatial support

Description

readAreaInfo will read a text file with observations and descriptions of data with a spatial support.

Usage


readAreaInfo(fname = "ainfo.txt", id = "id",
                          iobs = "iobs",obs="obs",unc="unc",filenames="filenames",
                          sep = "\t", debug.level = 1, moreCols = list(NULL)) 

Arguments

fname 	 name of file with areal information

id 	 name of column with observation id

iobs 	 name of column with number of observations

obs 	 name of column with observations

unc 	 name of column with possible uncertainty of observation

filenames 	 name of column with filenames of areas if different names than id should be used.

sep 	 separator in csv-file

debug.level 	 used for giving additional output

moreCols 	 name of other column names the user wants included in ainfo

Details

The function is of particular use when data are not available as shape-files, or when the observations are not part of the shape-files The simplest way to read the data in that case is through readShapePoly in the maptools-package or readOGR in the rgdal-package. See also rtopIntroduction.

Value

SpatialPointDataFrame with information about observations and/or predictionLocations.

Author(s)

Jon Olav Skoien

[Package rtop version 0.3-3 Index]