readAreas(rtop)	R Documentation

help file for creating SpatialPolygonsDataFrame with observations and/or predictionLocations of data with a spatial support

Description

readAreas will read area-files, add observations and convert the result to SpatialPolygonsDataFrame

Usage


readAreas(object, adir=".",ftype = "xy",projection = NA, ...) 

Arguments

object 	 either name of file with areal information or SpatialPointsDataFrame with observations

adir 	 directory where the files with areal information are to be found

ftype 	 type of file, possibilities are:

xy

referring to x- and y-coordinates of boundaries

shp

areas in individual shape files

shps

areas are in one individual shape file

projection 	 add projection to the object if input is boundary-files

... 	 further parameters to be passed to readAreaInfo

Details

If object is a file name, readAreaInfo will be called. If it is a SpatialPointsDataFrame with observations and/or predictionLocations, the function will read areal data from files according to the ID associated with each observation/predictionLocation.

The function is of particular use when data are not available as shape-files, or when the observations are not part of the shape-files The simplest way to read the data in that case is through readShapePoly in the maptools-package or readOGR in the rgdal-package. See also rtopIntroduction.

Value

The function creates a SpatialPolygonsDataFrame of observations and/or predictionLocations, depending on the information given in object.

Author(s)

Jon Olav Skoien

[Package rtop version 0.3-3 Index]