rtopFitVariogram(rtop)	R Documentation

Fit variogram model to sample variogram of data with spatial or line support

Description

rtopFitVariogram will fit a variogram model to the estimated binned variogram or cloud variogram of data with an areal or a line support.

Usage


## S3 method for class 'rtop':
rtopFitVariogram(object, ...) 
## S3 method for class 'SpatialPolygonsDataFrame':
rtopFitVariogram(object, params=list(), ...) 
## S3 method for class 'SpatialLinesDataFrame':
rtopFitVariogram(object, params=list(), ...) 
## S3 method for class 'SpatialPointsDataFrame':
rtopFitVariogram(object, params=list(), ...) 
## S3 method for class 'rtopVariogram':
rtopFitVariogram(object, observations, dists = NULL, params=list(), mr = FALSE, ...) 
## S3 method for class 'rtopVariogramCloud':
rtopFitVariogram(object, observations, dists = NULL, params=list(), mr = FALSE, ...) 

Arguments

object 	 object of class rtopVariogram or rtopVariogramCloud, or an object with class rtop that includes the sample variograms.
The object can also be of class SpatialPolygonsDataFrame, SpatialLinesDataFrame, or SpatialPointsDataFrame with observations. If object is a SpatialPointsDataFrame, it must have a column with name area (for areal observations) or length (for line observations).

observations 	 the observations, passed as a Spatial*DataFrame object, if object is an rtopVariogram or rtopVariogramCloud

params 	 a set of parameters, used to modify the standard parameters for the rtop package, set in getRtopParams.

dists 	 either a matrix with geostatistical distances (created by a call to the function geoDist or a list with the areas or lines discretized (from a call to rtopDiscAreas.

mr 	 logical; defining whether the function should return a list with discretized areas and geostatistical distances, even if it was not called with an rtop-object as argument.

... 	 Other parameters to functions called from rtopFitVarigoram

Value

The function creates an object with the fitted variogram Model (variogramModel) and a data.frame (varFit) with the differences between the sample semivariances and the regularized semivariances. If mr = TRUE, the function also returns other objects (discretized areas and geostatistical distances, if created) as a part of the returned object. If the function is called with an rtop-object as argument, it will return an rtop-object with variogramModel and varFit added to the object, in addition to other objects created.

Note

There are several options for fitting of the variogramModel, where the parameters can be set in params, which is a list of parameters for modification of the standard parameters of the rtop-package given in a call to getRtopParams. The first choice is between individual fitting and binned fitting. This is based on the type of variogram submitted, individual fitting is done if a cloud variogram (of class rtopVariogramCloud) is passed as argument, and binned fitting if the submitted variogram is of class rtopVariogram. If the function is called with an object of class rtop, having both variogram and variogramCloud among its arguments, the variogram model is fitted to the variogram which is consistent with the parameter cloud.

Author(s)

Jon Olav Skoien

References

http://www.intamap.org/

[Package rtop version 0.3-3 Index]