rtopKrige(rtop)	R Documentation

Spatial interpolation of data with spatial or line support

Description

rtopKrige perform spatial interpolation or cross validation of data with areal or a line support.

Usage


## S3 method for class 'rtop':
rtopKrige(object, cv, ...) 
## S3 method for class 'SpatialPolygonsDataFrame':
rtopKrige(object, predictionLocations = NULL,
    varMatObs, varMatPred, varMatPredObs, varMat, params = list(), formulaString, 
    cv, nmax, maxdist, wlim, debug.level, areas, ...) 

Arguments

object 	 object of class rtop or SpatialPolygonsDataFrame

predictionLocations 	 SpatialPolygons or SpatialPolygonsDataFrame with prediction locations. NULL if cross validation is to be performed.

varMatObs 	 covariance matrix of observations, where diagonal must consist of internal variance, typically generated from call to varMat

varMatPred 	 covariance matrix between prediction locations, where diagonal must consist of internal variance, typically generated from call to varMat. Can also be passed as an array of the diagonal.

varMatPredObs 	 covariance matrix between observation locations and prediction locations, typically generated from call to varMat

varMat 	 list of all covariance matrices above

params 	 a set of parameters, used to modify the standard parameters for the rtop package, set in getRtopParams. Additionally, it is possible overrule some of the parameters in object$params by passing them as separate arguments.

formulaString

cv

nmax

maxdist

wlim

debug.level

areas 	 array of area numbers, if only a limited number of areas are to be interpolated/crossvalidated

... 	 arguments to rtopKrige.rtop to be passed to rtopKrige.SpatialPolygonsDataFrame. In rtopKrige.SpatialPolygonsDataFrame, possible other parameters for getRtopParams, if these have not been set.

Details

This function is the interpolation routine of the rtop-package. The simplest way of calling the function is with an rtop-object that contains the fitted variogram model and all the other necessary data (see createRtopObject or rtopIntroduction).

The function will, if called with covariance matrices between observations and between observations and prediction locations, use these for the interpolation. If the function is called without these matrices, varMat will be called to create them. These matrices can therefore be reused if necessary, an advantage as it is computationally expensive to create them.

Value

If called with SpatialPolygonsDataFrame, the function returns a SpatialPolygonsDataFrame with predictions, either at the locations defined in predictionLocations, or as leave-one-out cross-validation predicitons at the same locations as in object if cv = TRUE
If called with an rtop-object, the function returns the same object with the predictions added to the object.

Author(s)

Jon Olav Skoien

[Package rtop version 0.3-8 Index]