rtopVariogram(rtop)	R Documentation

create variogram for data with spatial support

Description

rtopVariogram will create binned variogram or cloud variogram of data with an areal or a line support.

Usage


## S3 method for class 'rtop':
rtopVariogram(object, ...) 
## S3 method for class 'SpatialPolygonsDataFrame':
rtopVariogram(object, ...) 
## S3 method for class 'SpatialLinesDataFrame':
rtopVariogram(object, ...) 
## S3 method for class 'SpatialPointsDataFrame':
rtopVariogram(object, formulaString, params=list(), cloud, abins, dbins, ...) 

Arguments

object 	 object of class rtop (see rtopIntroduction) or a SpatialPolygonsDataFrame, SpatialLinesDataFrame, or SpatialPointsDataFrame with information about observations. If object is a SpatialPointsDataFrame, it must have a column with name area (for areal observations) or length (for line observations).

formulaString 	 formula that defines the dependent variable as a linear model of independent variables; suppose the dependent variable has name z, for ordinary and simple kriging use the formula z~1; for universal kriging, suppose z is linearly dependent on x and y, use the formula z~x+y. The formulaString defaults to "value~1" if value is a part of the data set. If not, the first column of the data set is used.

params 	 a set of parameters, used to modify the standard parameters for the rtop package, set in getRtopParams.

cloud 	 logical; if TRUE, calculate the semivariogram cloud, can be used to overrule the cloud parameter in params.

abins 	 possibility to set areal bins (not yet implemented)

dbins 	 possibility to set distance bins (not yet implemented)

... 	 parameters to other functions called, e.g. gstat's variogram-function and to rtopVariogram.SpatialPointsDataFrame when the method is called with an object of a different class

Value

The function creates a variogram, either of type rtopVariogram or rtopVariogramCloud. This variogram is based on the variogram function from gstat, but has additional information about the spatial size or length of the observations. An rtop-object with the variogram added is returned if the function is called with an rtop-object as argument.

Note

The variogram cloud is similar to the variogram cloud from gstat, with the area/length added to the resulting data.frame. The binned variogram is also based on the area or length, in addition to the distance between observations. The bins equally distanced in the log10-space of the distances and areas (lengths). The size of the bins is decided from the parameters amul and dmul, defining the number of bins per order of magnitude (1:10, 10:100, and so on).

The distances between areas/lines are at this function based on the centre of gravity.

Author(s)

Jon Olav Skoien

[Package rtop version 0.3-3 Index]