varMat(rtop)	R Documentation

create a semivariogram matrix between a set of locations or semivariogram matrices between and within two sets of locations

Description

varMat will create a semivariogram matrix between all the supports in a set of locations (observations or prediction locations) or semivariogram matrices between all the supports in one or two sets of locations, and also between them.

Usage


## S3 method for class 'rtop':
varMat(object, ...) 
## S3 method for class 'SpatialPolygonsDataFrame':
varMat(object,object2 = NULL,...) 
## S3 method for class 'SpatialPolygons':
varMat(object,object2 = NULL,variogramModel,
     params = list(), overlapObs, overlapPredObs, ...) 
## S3 method for class 'matrix':
varMat(object, variogramModel,  ...) 
## S3 method for class 'list':
varMat(object, object2=NULL, coor1, coor2, maxdist = Inf, 
              variogramModel, diag=FALSE, sub1, sub2, ...) 

Arguments

object 	 object of class rtop (see rtopIntroduction) or a SpatialPolygonsDataFrame, SpatialPolygons, or matrix or list with information about the observation supports.

object2 	 information about support of prediction locations

variogramModel 	 variogramModel to be used in calculation of the semivariogram matrix (matrices)

params 	 typical parameters to modify from the default parameters of the rtop-package (or modifications of the previously set parameters). Typical parameters to modify for this function:

rresol = 100

miminum number of discretization points, in call to rtopDiscAreas if necessary

rstype = "rtop"

sampling type from areas, in call to rtopDiscAreas if necessary

geoDistPred = FALSE

use geostatistical distance for semivariogram matrices

geoDist

parameter to set jointly geoDistEst = geoDistPred = geoDist

overlapObs 	 matrix with observations that overlap each other

overlapPredObs 	 matrix with observations and predictionLocations that overlap each other

coor1 	 coordinates of centroids of object

coor2 	 coordinates of centre-of-gravity of object2

maxdist 	 maximum distance between areas for inclusion in semivariogrma matrix

diag 	 logical; if TRUE only the semivariogram values along the diagonal will be calculated, typical for semivariogram matrix of prediction locations

sub1 	 semivariogram array for subtraction of inner variances of areas

sub2 	 semivariogram array for subtraction of inner variances of areas

... 	 Other parameters to params

Author(s)

Jon Olav Skoien

[Package rtop version 0.3-3 Index]